np.array を np.fromiter に変更

+++ example/pmx01.model.py
@@ -10,10 +10,11 @@ sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
 import numpy as np
 import OpenGL.GL as gl
 import wx
-from mlib.math import MMatrix4x4, MQuaternion, MVector3D
-from mlib.pmx.pmx_reader import PmxReader
 from OpenGL.GL import shaders
 from wx import glcanvas
+
+from mlib.math import MMatrix4x4, MQuaternion, MVector3D
+from mlib.pmx.pmx_reader import PmxReader
 
 vertex_shader = """
 # version 330
@@ -66,7 +67,7 @@ class Geometries:
         #         0.0,
         #         1.0,
         #     ],
-        #     dtype=np.float32,
+        #     dtype=np.float64,
         # )
 
         # self.vao_triangle = gl.glGenVertexArrays(1)
@@ -118,7 +119,7 @@ class Geometries:
         #         1.0,
         #         1.0,
         #     ],
-        #     dtype=np.float32,
+        #     dtype=np.float64,
         # )
 
         # self.vao_quad = gl.glGenVertexArrays(1)
@@ -168,11 +169,11 @@ class Geometries:
                             ).vector
                             for vidx in self.model.faces[face_index].vertices
                         ],
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
             prev_face_count += face_count
-        self.vertices = np.array(vertex_position_list, dtype=np.float32)
+        self.vertices = np.array(vertex_position_list, dtype=np.float64)
 
         self.vao_vertices = gl.glGenVertexArrays(1)
         gl.glBindVertexArray(self.vao_vertices)
