+++ example/pmx02.hide.py
@@ -11,10 +11,11 @@ import numpy as np
 import OpenGL.GL as gl
 import OpenGL.GLU as glu
 import wx
-from mlib.math import MMatrix4x4, MQuaternion, MVector3D
-from mlib.pmx.pmx_reader import PmxReader
 from OpenGL.GL import shaders
 from wx import glcanvas
+
+from mlib.math import MMatrix4x4, MQuaternion, MVector3D
+from mlib.pmx.pmx_reader import PmxReader
 
 vertex_shader = """
 # version 330
@@ -71,11 +72,11 @@ class Geometries:
                             ).vector
                             for vidx in self.model.faces[face_index].vertices
                         ],
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
             prev_face_count += face_count
-        self.vertices = np.array(vertex_position_list, dtype=np.float32)
+        self.vertices = np.array(vertex_position_list, dtype=np.float64)
 
         self.vao_vertices = gl.glGenVertexArrays(1)
         gl.glBindVertexArray(self.vao_vertices)
