+++ example/pmx06.tex.py
@@ -141,7 +141,7 @@ class Geometries:
                             ).vector
                             for vidx in self.model.faces[face_index].vertices
                         ],
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 vertex_normal_list.append(
@@ -150,7 +150,7 @@ class Geometries:
                             self.model.vertices[vidx].normal.vector
                             for vidx in self.model.faces[face_index].vertices
                         ],
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 vertex_uv_list.append(
@@ -159,7 +159,7 @@ class Geometries:
                             self.model.vertices[vidx].uv.vector
                             for vidx in self.model.faces[face_index].vertices
                         ],
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 texture_index_list.append(
@@ -171,9 +171,9 @@ class Geometries:
                                     len(self.model.faces[face_index].vertices)
                                 )
                             ],
-                            dtype=np.float32,
+                            dtype=np.float64,
                         ),
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 diffuse_color_list.append(
@@ -185,9 +185,9 @@ class Geometries:
                                     len(self.model.faces[face_index].vertices)
                                 )
                             ],
-                            dtype=np.float32,
+                            dtype=np.float64,
                         ),
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 specular_color_list.append(
@@ -204,9 +204,9 @@ class Geometries:
                                     len(self.model.faces[face_index].vertices)
                                 )
                             ],
-                            dtype=np.float32,
+                            dtype=np.float64,
                         ),
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 ambient_color_list.append(
@@ -222,19 +222,19 @@ class Geometries:
                                     len(self.model.faces[face_index].vertices)
                                 )
                             ],
-                            dtype=np.float32,
+                            dtype=np.float64,
                         ),
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
             prev_face_count += face_count
-        self.vertices = np.array(vertex_position_list, dtype=np.float32)
-        self.normals = np.array(vertex_normal_list, dtype=np.float32)
-        self.uvs = np.array(vertex_uv_list, dtype=np.float32)
-        self.textures = np.array(texture_index_list, dtype=np.float32)
-        self.diffuses = np.array(diffuse_color_list, dtype=np.float32)
-        self.speculars = np.array(specular_color_list, dtype=np.float32)
-        self.ambients = np.array(ambient_color_list, dtype=np.float32)
+        self.vertices = np.array(vertex_position_list, dtype=np.float64)
+        self.normals = np.array(vertex_normal_list, dtype=np.float64)
+        self.uvs = np.array(vertex_uv_list, dtype=np.float64)
+        self.textures = np.array(texture_index_list, dtype=np.float64)
+        self.diffuses = np.array(diffuse_color_list, dtype=np.float64)
+        self.speculars = np.array(specular_color_list, dtype=np.float64)
+        self.ambients = np.array(ambient_color_list, dtype=np.float64)
         self.texture_indecies = list(range(len(self.model.textures)))
 
         # ---------------------
