+++ example/pmx07.440.py
@@ -138,7 +138,7 @@ class Geometries:
                             ).vector
                             for vidx in self.model.faces[face_index].vertices
                         ],
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 vertex_normal_list.append(
@@ -147,7 +147,7 @@ class Geometries:
                             self.model.vertices[vidx].normal.vector
                             for vidx in self.model.faces[face_index].vertices
                         ],
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 vertex_uv_list.append(
@@ -156,7 +156,7 @@ class Geometries:
                             self.model.vertices[vidx].uv.vector
                             for vidx in self.model.faces[face_index].vertices
                         ],
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 texture_index_list.append(
@@ -168,9 +168,9 @@ class Geometries:
                                     len(self.model.faces[face_index].vertices)
                                 )
                             ],
-                            dtype=np.float32,
+                            dtype=np.float64,
                         ),
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 diffuse_color_list.append(
@@ -182,9 +182,9 @@ class Geometries:
                                     len(self.model.faces[face_index].vertices)
                                 )
                             ],
-                            dtype=np.float32,
+                            dtype=np.float64,
                         ),
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 specular_color_list.append(
@@ -201,9 +201,9 @@ class Geometries:
                                     len(self.model.faces[face_index].vertices)
                                 )
                             ],
-                            dtype=np.float32,
+                            dtype=np.float64,
                         ),
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
                 ambient_color_list.append(
@@ -219,19 +219,19 @@ class Geometries:
                                     len(self.model.faces[face_index].vertices)
                                 )
                             ],
-                            dtype=np.float32,
+                            dtype=np.float64,
                         ),
-                        dtype=np.float32,
+                        dtype=np.float64,
                     )
                 )
             prev_face_count += face_count
-        self.vertices = np.array(vertex_position_list, dtype=np.float32)
-        self.normals = np.array(vertex_normal_list, dtype=np.float32)
-        self.uvs = np.array(vertex_uv_list, dtype=np.float32)
-        self.textures = np.array(texture_index_list, dtype=np.float32)
-        self.diffuses = np.array(diffuse_color_list, dtype=np.float32)
-        self.speculars = np.array(specular_color_list, dtype=np.float32)
-        self.ambients = np.array(ambient_color_list, dtype=np.float32)
+        self.vertices = np.array(vertex_position_list, dtype=np.float64)
+        self.normals = np.array(vertex_normal_list, dtype=np.float64)
+        self.uvs = np.array(vertex_uv_list, dtype=np.float64)
+        self.textures = np.array(texture_index_list, dtype=np.float64)
+        self.diffuses = np.array(diffuse_color_list, dtype=np.float64)
+        self.speculars = np.array(specular_color_list, dtype=np.float64)
+        self.ambients = np.array(ambient_color_list, dtype=np.float64)
         self.texture_indecies = list(range(len(self.model.textures)))
 
         # ---------------------
